requirements.txt
joblib
jupyter
lightgbm
matplotlib
mysql-connector-python
nbstripout
numpy
pandas
python-dateutil
python-dotenv
pyyaml
scikit-learn
scipy
seaborn
snowflake-connector-python
statsmodels
timezonefinder
redis